requirements.txt
# Core deep learning
torch==2.6.0
torchvision==0.21.0

# Medical imaging
monai==1.5.0
lighter-zoo==0.1.3
nibabel==5.3.2

# Model backbone
timm==1.0.15

# Data & numerics
numpy==2.2.6
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2
scikit-image==0.25.2

# Visualization
matplotlib==3.10.7
seaborn==0.13.2
umap-learn==0.5.11

# Experiment tracking
wandb==0.19.7
tensorboard==2.19.0

# Utilities
tqdm==4.67.1
